requirements.txt
grpcio>=1.59.0
grpcio-tools>=1.59.0
ccxt>=4.1.13
protobuf>=4.25.1 